+++ print_octal.c
@@ -78,7 +78,7 @@ int print_short_octal(va_list args)
 	short int r;
 
 	/* va_arg calls the argument provided and passes it as unsigned int */
-	num = va_arg(args, unsigned short int);
+	num = va_arg(args, unsigned int);
 
 	if (num == 0)
 	{
