+++ print_unsigned_hex.c
@@ -152,7 +152,7 @@ int print_short_unsigned_hex(va_list args)
 	short int r;
 
 	/* va_arg calls the argument provided and passes it as unsigned int */
-	num = va_arg(args, unsigned short int);
+	num = va_arg(args, unsigned int);
 
 	if (num == 0)
 	{
